requirements.txt
discord.py\ngoogle-generativeai\npython-dotenv
